adding claim process first part

+++ database/migrations/2020_10_13_052809_add_person_relationship_user_table.php
@@ -14,7 +14,7 @@ class AddPersonRelationshipUserTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint  $table) {
-           $table->unsignedInteger('person_id')->nullable(true);
+           $table->unsignedBigInteger('person_id')->nullable(true);
 
            $table->foreign('person_id')
                ->references('id')
